fix: move local flags back to local model
